ChainDB: split off all path/candidate related code from the VolDB
This is in preparation of unifying the VolDB with the VolatileDB.
