+++ ouroboros-consensus-test/ouroboros-consensus-test.cabal
@@ -207,8 +207,8 @@ test-suite test-storage
                        Test.Ouroboros.Storage.ChainDB.GcSchedule
                        Test.Ouroboros.Storage.ChainDB.Model
                        Test.Ouroboros.Storage.ChainDB.Model.Test
+                       Test.Ouroboros.Storage.ChainDB.Paths
                        Test.Ouroboros.Storage.ChainDB.StateMachine
-                       Test.Ouroboros.Storage.ChainDB.VolDB
                        Test.Ouroboros.Storage.FS
                        Test.Ouroboros.Storage.FS.StateMachine
                        Test.Ouroboros.Storage.ImmutableDB
