+++ ouroboros-consensus-test/test-storage/Test/Ouroboros/Storage/ChainDB.hs
@@ -7,14 +7,14 @@ import           Test.Tasty
 import qualified Test.Ouroboros.Storage.ChainDB.GcSchedule as GcSchedule
 import qualified Test.Ouroboros.Storage.ChainDB.Iterator as Iterator
 import qualified Test.Ouroboros.Storage.ChainDB.Model.Test as Model
+import qualified Test.Ouroboros.Storage.ChainDB.Paths as Paths
 import qualified Test.Ouroboros.Storage.ChainDB.StateMachine as StateMachine
-import qualified Test.Ouroboros.Storage.ChainDB.VolDB as VolDB
 
 tests :: TestTree
 tests = testGroup "ChainDB" [
       Iterator.tests
     , GcSchedule.tests
     , Model.tests
+    , Paths.tests
     , StateMachine.tests
-    , VolDB.tests
     ]
