+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -184,6 +184,7 @@ library
                        Ouroboros.Consensus.Storage.ChainDB.Impl.ImmDB
                        Ouroboros.Consensus.Storage.ChainDB.Impl.Iterator
                        Ouroboros.Consensus.Storage.ChainDB.Impl.LgrDB
+                       Ouroboros.Consensus.Storage.ChainDB.Impl.Paths
                        Ouroboros.Consensus.Storage.ChainDB.Impl.Query
                        Ouroboros.Consensus.Storage.ChainDB.Impl.Reader
                        Ouroboros.Consensus.Storage.ChainDB.Impl.Types
